refactor: 길이 기반 전략 수정
