+++ src/main/java/chess/domain/piece/Knight.java
@@ -2,12 +2,23 @@ package chess.domain.piece;
 
 import chess.domain.Color;
 import chess.domain.piece.strategy.LengthBasedMovingStrategy;
+import chess.domain.position.Position;
+import java.util.List;
 
 public class Knight extends Piece {
 
-    private static final int MOVABLE_LENGTH = 5;
+    private static final LengthBasedMovingStrategy MOVING_STRATEGY = new LengthBasedMovingStrategy(number -> number == 5);
 
     public Knight(Color color) {
-        super(PieceType.KNIGHT, color, new LengthBasedMovingStrategy(number -> number != MOVABLE_LENGTH));
+        super(PieceType.KNIGHT, color, null);
+    }
+
+    @Override
+    public void validateMove(List<List<Piece>> board, Position source, Position target) {
+        if (MOVING_STRATEGY.canMove(board, source, target)) {
+            return;
+        }
+
+        throw new IllegalArgumentException();
     }
 }
