--- src/main/java/chess/domain/piece/strategy/king/KingMovingStrategy.java
@@ -1,11 +1,0 @@
-package chess.domain.piece.strategy.king;
-
-import chess.domain.piece.Piece;
-import chess.domain.position.Position;
-import java.util.List;
-
-@FunctionalInterface
-public interface KingMovingStrategy {
-
-    boolean canMove(List<List<Piece>> board, Position source, Position target);
-}
